refactor(Conversations): Move model to typescript
